flask/pyproject.toml
[project]
name = "wca-grid"

version = "1.0.0"

dependencies = [
    "flask",
    "flask-cors",
    "psycopg2-binary",
    "python-dotenv"
]

[tool.setuptools]
py-modules = []